+++ src/pages/post-write/ui/PostWritePage.tsx
@@ -1,8 +1,6 @@
-import { Button, useDisclosure, Box, Flex, VStack, Divider } from '@chakra-ui/react';
+import { useDisclosure, VStack } from '@chakra-ui/react';
 
-import { ArrowLeft } from 'lucide-react';
-
-import { PostTitle, PostTag, PostContent } from '@shared/components';
+import { PostTitle, PostTag, PostContent, PostButtons } from '@shared/components';
 
 import { PostModal } from '@widgets/modals';
 
@@ -10,49 +8,18 @@ export const PostWritePage = () => {
   const { isOpen, onOpen, onClose } = useDisclosure();
 
   return (
-    <Flex w='full' justifyContent='center'>
-      <VStack w='full' py='20px' gap='0'>
-        <PostTitle />
-        <Box w='full' h='14px' bg='white'>
-          <Divider
-            w={{ base: '145px', md: '215px' }}
-            h='8px'
-            bg='black'
-            ml={{ base: '10px', md: '50px' }}
-          />
-        </Box>
-        <PostTag />
-        <PostContent />
-        <Flex
-          w='full'
-          h='50px'
-          px={{ base: '10px', sm: '30px' }}
-          mt='5px'
-          bg='white'
-          align='center'
-          justify='space-between'
-        >
-          <Button
-            variant='ghost'
-            colorScheme='gray'
-            fontSize='md'
-            leftIcon={<ArrowLeft />}
-            _hover={{ bg: 'transparent' }}
-          >
-            나가기
-          </Button>
-          <Button variant='solid' colorScheme='blue' py={4} fontSize='md' onClick={onOpen}>
-            작성완료
-          </Button>
-        </Flex>
-        <PostModal
-          title='백준 1004번 풀이'
-          isOpen={isOpen}
-          onClose={onClose}
-          buttonTitle='출간하기'
-          postType='create'
-        />
-      </VStack>
-    </Flex>
+    <VStack w='full' py='20px' gap='0'>
+      <PostTitle />
+      <PostTag />
+      <PostContent />
+      <PostButtons buttonText='작성완료' onClick={onOpen} />
+      <PostModal
+        title='백준 1004번 풀이'
+        isOpen={isOpen}
+        onClose={onClose}
+        buttonTitle='출간하기'
+        postType='create'
+      />
+    </VStack>
   );
 };
