+++ src/shared/components/post-form/index.ts
@@ -1,3 +1,4 @@
-export * from './PostTitle';
+export * from './post-title';
 export * from './post-tag';
 export * from './post-content';
+export * from './post-button';
