+++ src/shared/components/post-form/post-button/index.ts
@@ -1,0 +1,1 @@
+export * from './ui';
