+++ src/shared/components/post-form/post-button/ui/index.ts
@@ -1,0 +1,1 @@
+export { PostButtons } from './PostButtons';
