+++ src/shared/components/post-form/post-title/index.ts
@@ -1,0 +1,1 @@
+export * from './ui';
