+++ src/shared/components/post-form/post-title/ui/index.ts
@@ -1,0 +1,1 @@
+export { PostTitle } from './PostTitle';
